LPS-76668 Avoid clobbering the file as a separate import is happening with the older version, and prevent reading from an incomplete file

+++ portal-impl/src/com/liferay/portlet/sites/util/SitesImpl.java
@@ -1878,12 +1878,12 @@ public class SitesImpl implements Sites {
 		if (!importData) {
 			cacheFile = new File(
 				StringBundler.concat(
-					_TEMP_DIR, layoutSetPrototype.getUuid(), ".lar"));
+					_TEMP_DIR, layoutSetPrototype.getUuid(), StringPool.DASH,
+					String.valueOf(layoutSetPrototype.getMvccVersion()),
+					".lar"));
 
-			if (cacheFile.exists()) {
-				Date modifiedDate = layoutSetPrototype.getModifiedDate();
-
-				if (cacheFile.lastModified() >= modifiedDate.getTime()) {
+			synchronized (this) {
+				if (cacheFile.exists()) {
 					if (_log.isDebugEnabled()) {
 						_log.debug(
 							"Using cached layout set prototype LAR file " +
@@ -1946,13 +1946,17 @@ public class SitesImpl implements Sites {
 
 		if (!importData && newFile) {
 			try {
-				FileUtil.copyFile(file, cacheFile);
+				synchronized (this) {
+					if (!cacheFile.exists()) {
+						FileUtil.copyFile(file, cacheFile);
 
-				if (_log.isDebugEnabled()) {
-					_log.debug(
-						StringBundler.concat(
-							"Copied ", file.getAbsolutePath(), " to ",
-							cacheFile.getAbsolutePath()));
+						if (_log.isDebugEnabled()) {
+							_log.debug(
+								StringBundler.concat(
+									"Copied ", file.getAbsolutePath(), " to ",
+									cacheFile.getAbsolutePath()));
+						}
+					}
 				}
 			}
 			catch (Exception e) {
